tests/wpt/tests/html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-reload-navigation-reload.html

<!DOCTYPE html>
<title>Reloading iframe loading='lazy' before it is loaded: location.reload</title>
<iframe src="support/blank.htm?src" loading="lazy" hidden></iframe>
<script>
const iframe = document.querySelector('iframe');
const iframeLoaded = new Promise(resolve => {
  iframe.onload = resolve;
});
// Should have no effect on lazy-loading
iframe.contentWindow.navigation.reload();
iframe.hidden = false;
</script>
<!-- Loading testharness.js here is intentional to reproduce a bug in WebKit. -->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({single_test: true});
iframeLoaded.then(() => {
  // Need a timeout to detect failure when there are two navigations.
  step_timeout(() => {
    assert_true("navigation" in window, "Navigation API is supported");
    assert_equals(iframe.contentWindow.location.href, new URL("support/blank.htm?src", location.href).href);
    done();
  }, 1000);
});
</script>
