third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-reload-navigation-reload-expected.txt

This is a testharness.js-based test.
[FAIL] Reloading iframe loading='lazy' before it is loaded: location.reload
  Uncaught Error: assert_equals: expected "http://web-platform.test:8001/html/semantics/embedded-content/the-iframe-element/support/blank.htm?src" but got "about:blank"
Harness: the test ran to completion.
